Merged my entries from 2000.  Obviously I forgot to commit the Changelog to the CVS when moving the files from the local PRCS repository to CVS.
Fixed copyright entries from "Free Software Foundation, Inc" to
"Werner Koch".  It was originally intended to assign the code to the
FSF but this actually never happened and frankly there has been no
real release as the code lived his life for most time merely as a test
bed for libgcrypt.